Add files via upload
Created a Restaurant webpage using HTML and cascading style sheets. The name of the restaurant
given by us is AROMA which is displayed at the top left of the webpage. It also features five 
options at the center top Home, Menu, Chef, About, and Contact. It also displays a number on the
top right of the page. These are buttons and the top of the website looks like a menu bar. There 
is an image right below this menu bar. There is a text that mentions the restaurant's name, a 
oneliner, and a menu button option. These buttons turn blue when the cursor is placed over them.
When the menu button is clicked whether from the center top or from the option available at the 
middle of the image it scrolls down into the menu. At this part, one will be able to view the menu
provided by the restaurant. It contains 5 sets of breakfast, lunch, dinner, dessert, and drinks. 
If the chef button is clicked it further scrolls downs to a section that shows the chefs working 
for the restaurant. If the about button is clicked it further scrolls down and shows a para about
the "About us" of the restaurant and there is an image near it. And at last, when the Contact 
button is clicked it scrolls down to the bottom of the webpage, and at the left side, it shows the
location of the restaurant, mail, and phone number. Below it there are also links for social 
accounts. The mail when clicked opens the mail in one's device the same happens if you click the
phone number. There is also a box available at the bottom right of the webpage.